refactor: autoItemPickup & carefulBreak
This change uses the same pattern as instamineable
